+++ packages/providers/lib.esm/alchemy-provider.js
@@ -30,6 +30,9 @@ export class AlchemyProvider extends UrlJsonRpcProvider {
             case "rinkeby":
                 host = "eth-rinkeby.alchemyapi.io/jsonrpc/";
                 break;
+            case "goerli":
+                host = "eth-goerli.alchemyapi.io/jsonrpc/";
+                break;
             case "kovan":
                 host = "eth-kovan.alchemyapi.io/jsonrpc/";
                 break;