+++ packages/providers/lib.esm/base-provider.js
@@ -100,6 +100,7 @@ function getTime() {
 /**
  *  EventType
  *   - "block"
+ *   - "poll"
  *   - "pending"
  *   - "error"
  *   - filter
@@ -148,7 +149,7 @@ export class Event {
         return filter;
     }
     pollable() {
-        return (this.tag.indexOf(":") >= 0 || this.tag === "block" || this.tag === "pending");
+        return (this.tag.indexOf(":") >= 0 || this.tag === "block" || this.tag === "pending" || this.tag === "poll");
     }
 }
 let defaultFormatter = null;
@@ -264,8 +265,11 @@ export class BaseProvider extends Provider {
             const runners = [];
             const blockNumber = yield this._getInternalBlockNumber(100 + this.pollingInterval / 2);
             this._setFastBlockNumber(blockNumber);
+            // Emit a poll event after we have the latest (fast) block number
+            this.emit("poll", pollId, blockNumber);
             // If the block has not changed, meh.
             if (blockNumber === this._lastBlockNumber) {
+                this.emit("didPoll", pollId);
                 return;
             }
             // First polling cycle, trigger a "block" events
@@ -358,22 +362,38 @@ export class BaseProvider extends Provider {
         return this.ready;
     }
     get blockNumber() {
-        return this._fastBlockNumber;
+        this._getInternalBlockNumber(100 + this.pollingInterval / 2).then((blockNumber) => {
+            this._setFastBlockNumber(blockNumber);
+        });
+        return (this._fastBlockNumber != null) ? this._fastBlockNumber : -1;
     }
     get polling() {
         return (this._poller != null);
     }
     set polling(value) {
-        setTimeout(() => {
-            if (value && !this._poller) {
-                this._poller = setInterval(this.poll.bind(this), this.pollingInterval);
-                this.poll();
+        if (value && !this._poller) {
+            this._poller = setInterval(this.poll.bind(this), this.pollingInterval);
+            if (!this._bootstrapPoll) {
+                this._bootstrapPoll = setTimeout(() => {
+                    this.poll();
+                    // We block additional polls until the polling interval
+                    // is done, to prevent overwhelming the poll function
+                    this._bootstrapPoll = setTimeout(() => {
+                        // If polling was disabled, something may require a poke
+                        // since starting the bootstrap poll and it was disabled
+                        if (!this._poller) {
+                            this.poll();
+                        }
+                        // Clear out the bootstrap so we can do another
+                        this._bootstrapPoll = null;
+                    }, this.pollingInterval);
+                }, 0);
             }
-            else if (!value && this._poller) {
-                clearInterval(this._poller);
-                this._poller = null;
-            }
-        }, 0);
+        }
+        else if (!value && this._poller) {
+            clearInterval(this._poller);
+            this._poller = null;
+        }
     }
     get pollingInterval() {
         return this._pollingInterval;