+++ packages/providers/lib/alchemy-provider.js
@@ -48,6 +48,9 @@ var AlchemyProvider = /** @class */ (function (_super) {
             case "rinkeby":
                 host = "eth-rinkeby.alchemyapi.io/jsonrpc/";
                 break;
+            case "goerli":
+                host = "eth-goerli.alchemyapi.io/jsonrpc/";
+                break;
             case "kovan":
                 host = "eth-kovan.alchemyapi.io/jsonrpc/";
                 break;