+++ packages/tests/lib/test-providers.js
@@ -406,12 +406,6 @@ function testProvider(providerName, networkName) {
             if (providerName === "getDefaultProvider") {
                 provider = ethers_1.ethers.getDefaultProvider();
             }
-            else if (providerName === "Web3Provider") {
-                //let infuraUrl = (new ethers.providers.InfuraProvider()).connection.url;
-                //provider = new ethers.providers.Web3Provider(new Web3HttpProvider(infuraUrl));
-                //throw new Error("skipping Web3 tests; archos error on npm install");
-                return;
-            }
             else {
                 provider = new (ethers_1.ethers.providers)[providerName]();
             }
@@ -419,12 +413,6 @@ function testProvider(providerName, networkName) {
         else {
             if (providerName === "getDefaultProvider") {
                 provider = ethers_1.ethers.getDefaultProvider(networkName);
-            }
-            else if (providerName === "Web3Provider") {
-                //let infuraUrl = (new ethers.providers.InfuraProvider(networkName)).connection.url;
-                //provider = new ethers.providers.Web3Provider(new Web3HttpProvider(infuraUrl), networkName);
-                //throw new Error("skipping Web3 tests; archos error on npm install");
-                return;
             }
             else {
                 provider = new (ethers_1.ethers.providers)[providerName](networkName);
@@ -563,7 +551,7 @@ function testProvider(providerName, networkName) {
         if (providerName === "CloudflareProvider") {
             return;
         }
-        if (networkName === "goerli" && providerName === "AlchemyProvider") {
+        if (providerName === "Web3Provider") {
             return;
         }
         if ((networkName !== "homestead" && networkName !== "default") && providerName === "CloudflareProvider") {